fixing return value and adding docstring, refs #22
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@22070 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/common/TplConvert.h
@@ -191,7 +191,7 @@ public:
     }
     //@}
 
-    /// @name Conversion to basic type long
+    /// @name Conversion to basic type long integer (at least 64 bit)
     //@{
     /// @brief converts a char-type array into the long value described by it
     /// @throw an EmptyData - exception if the given string is empty
@@ -226,7 +226,10 @@ public:
         return ret * sgn;
     }
 
-    static int _str2long(const std::string& sData) {
+    /// @brief converts a string into the long value described by it by calling the char-type converter, which
+    /// @throw an EmptyData - exception if the given string is empty
+    /// @throw NumberFormatException - exception when the string does not contain a long integer
+    static long long int _str2long(const std::string& sData) {
         return _2long(sData.c_str());
     }
